#208 move pruning logic into AiPointsExpansionService
